Revert "Move examples to seperate crate, resolve circular dependancy"
This reverts commit 5e0bac90e0663a08bcdb105767bd229036b87fc6.
