--- crates/examples/Cargo.toml
@@ -1,129 +1,0 @@
-[package]
-authors = ["Espresso Systems <hello@espressosys.com>"]
-description = "HotShot Examples and binaries"
-edition = "2021"
-name = "hotshot-examples"
-readme = "README.md"
-version = "0.3.3"
-rust-version = "1.65.0"
-
-[features]
-default = ["docs", "doc-images"]
-
-# Features required for binaries
-bin-orchestrator = ["clap"]
-
-# Build the extended documentation
-docs = []
-doc-images = []
-hotshot-testing = []
-randomized-leader-election = []
-
-# libp2p
-[[example]]
-name = "validator-libp2p"
-path = "libp2p/validator.rs"
-
-[[example]]
-name = "multi-validator-libp2p"
-path = "libp2p/multi-validator.rs"
-
-[[example]]
-name = "orchestrator-libp2p"
-path = "libp2p/orchestrator.rs"
-
-[[example]]
-name = "all-libp2p"
-path = "libp2p/all.rs"
-
-# webserver
-[[example]]
-name = "webserver"
-path = "webserver/webserver.rs"
-
-[[example]]
-name = "orchestrator-webserver"
-path = "webserver/orchestrator.rs"
-
-[[example]]
-name = "validator-webserver"
-path = "webserver/validator.rs"
-
-[[example]]
-name = "multi-validator-webserver"
-path = "webserver/multi-validator.rs"
-
-[[example]]
-name = "multi-webserver"
-path = "webserver/multi-webserver.rs"
-
-[[example]]
-name = "all-webserver"
-path = "webserver/all.rs"
-
-# combined
-[[example]]
-name = "all-combined"
-path = "combined/all.rs"
-
-[[example]]
-name = "multi-validator-combined"
-path = "combined/multi-validator.rs"
-
-[[example]]
-name = "validator-combined"
-path = "combined/validator.rs"
-
-[[example]]
-name = "orchestrator-combined"
-path = "combined/orchestrator.rs"
-
-[dependencies]
-async-broadcast = { workspace = true }
-async-compatibility-layer = { workspace = true }
-async-lock = { workspace = true }
-async-trait = { workspace = true }
-bimap = "0.6.3"
-bincode = { workspace = true }
-clap = { version = "4.4", features = ["derive", "env"], optional = true }
-commit = { workspace = true }
-hotshot-constants = { path = "../constants" }
-custom_debug = { workspace = true }
-dashmap = "5.5.1"
-either = { workspace = true }
-embed-doc-image = "0.1.4"
-futures = { workspace = true }
-hotshot-web-server = { version = "0.1.1", path = "../web_server", default-features = false }
-hotshot-orchestrator = { version = "0.1.1", path = "../orchestrator", default-features = false }
-hotshot-types = { path = "../types", version = "0.1.0", default-features = false }
-hotshot-utils = { path = "../utils" }
-hotshot-task-impls = { path = "../task-impls", version = "0.1.0", default-features = false }
-libp2p-identity = { workspace = true }
-libp2p-networking = { workspace = true }
-rand = { workspace = true }
-serde = { workspace = true, features = ["rc"] }
-snafu = { workspace = true }
-surf-disco = { workspace = true }
-time = { workspace = true }
-derive_more = "0.99.17"
-portpicker = "0.1.1"
-lru = "0.12.2"
-hotshot-task = { path = "../task" }
-hotshot = { path = "../hotshot" }
-
-tracing = { workspace = true }
-
-[target.'cfg(all(async_executor_impl = "tokio"))'.dependencies]
-tokio = { workspace = true }
-[target.'cfg(all(async_executor_impl = "async-std"))'.dependencies]
-async-std = { workspace = true }
-
-[dev-dependencies]
-clap = { version = "4.4", features = ["derive", "env"] }
-toml = { workspace = true }
-blake3 = { workspace = true }
-local-ip-address = "0.5.7"
-hotshot-testing = { path = "../testing" }
-
-[lints]
-workspace = true
