+++ crates/hotshot/examples/combined/types.rs
