+++ crates/hotshot/examples/combined/validator.rs
