+++ crates/hotshot/examples/webserver/multi-webserver.rs
