+++ crates/hotshot/examples/webserver/types.rs
